requirements.txt
boto3==1.17.49
boto3-type-annotations==0.3.1
botocore==1.20.49
Cerberus==1.3.4
click==8.0.1
jmespath==0.10.0
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.3.6
six==1.15.0
urllib3==1.26.4
